feat: 시간에 대한 삭제 예외 처리 추가

+++ src/main/java/roomescape/service/TimeService.java
@@ -7,15 +7,18 @@ import org.springframework.stereotype.Service;
 import roomescape.dto.TimeRequest;
 import roomescape.dto.TimeResponse;
 import roomescape.model.ReservationTime;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.TimeRepository;
 
 @Service
 public class TimeService {
 
     private final TimeRepository timeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public TimeService(TimeRepository timeRepository) {
+    public TimeService(TimeRepository timeRepository, ReservationRepository reservationRepository) {
         this.timeRepository = timeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public List<TimeResponse> readAll() {
@@ -38,11 +41,16 @@ public class TimeService {
         } catch (EmptyResultDataAccessException e) {
             throw new IllegalArgumentException("삭제하고자 하는 시간 ID가 없습니다.");
         }
+
+        if (reservationRepository.existsByTimeId(id)) {
+            throw new IllegalArgumentException("현재 예약이 존재하는 시간은 삭제할 수 없습니다.");
+        }
+
         timeRepository.deleteById(id);
     }
 
     public TimeResponse register(TimeRequest timeRequest) {
-        if (timeRepository.existsByStartAt(timeRequest.startAt()) > 0) {
+        if (timeRepository.existsByStartAt(timeRequest.startAt())) {
             throw new IllegalArgumentException("이미 존재하는 시간입니다.");
         }
